+++ docs/CHANGELOG.md
@@ -19,6 +19,8 @@
 - added a new Lua function, `trx.rooms.flip_effect()`, to set the active flip effect with an optional timer (#4704)
 - added a new Lua catalog, `trx.catalog.flip_effects` for name-based flip effect catalog IDs
 - added new Lua catalogs for Lara states, Lara anims, music, and samples
+- added a new Lua module, `trx.camera`, with camera getters and `trx.camera.shake()`
+- added `num` to `trx.rooms.Room`
 - added support for cross-fades to the title screen
 - added visual previews of bar colors (Graphic Options → Bars)
 - added endless sprint (available previously via the `/restless` command) to the UI settings (Gameplay options → Mods → Endless sprint)